utils: add time_func.py and s1_utils.py
+ add utils.time_func.py:
   - move get_design_matrix4time_func() from objects.stack.py.
      - switch module import/usage in all related scripts: dem_error, gps, simulation, tsview
      - replace the old func definition with a pointer to the utils.time_func.py for backward compatibility.
   - move estimate_time_func() from timeseries2velocity.py
      - switch module import/usage in all related scripts: timeseries2velocity, tsview
   - set all get_design_matrix4*_func() as independent functions for more flexibility

+ utils.s1_utils:
   - add estimate_S1AB_bias()
   - add get_subswath_masks()

+ objects/gps.py: check site name during initialization for more informative error msg against typo in "view.py --ref-gps"

+ docs/api/module_hierarchy.md: add time_func and s1_utils.
